Update vasp datatype converter prototype
