+++ prototypes/datatype_converters/vasp_converter.py
@@ -1,0 +1,109 @@
+import json
+import sys
+import os
+
+from tqdm import tqdm
+
+from ..validator.schema_validator import Validator
+from ..parsers.ase_parser import parse_ase
+from ..utils.file_utils import find_files
+
+# VERSION 0.2.0
+
+# This is the datatype converter for VASP (OUTCAR).
+# Arguments:
+#   input_path (string): The file or directory where the data resides.
+#       NOTE: Do not hard-code the path to the data in the converter. The converter should be portable.
+#   metadata (string or dict): The path to the JSON dataset metadata file, or a dict or json.dumps string containing the dataset metadata.
+#   verbose (bool): Should the script print status messages to standard output? Default False.
+#       NOTE: The converter should have NO output if verbose is False, unless there is an error.
+def convert(input_path, metadata, verbose=False):
+    if verbose:
+        print("Begin converting")
+
+    # Collect the metadata
+    if type(metadata) is str:
+        try:
+            dataset_metadata = json.loads(metadata)
+        except Exception:
+            try:
+                with open(metadata, 'r') as metadata_file:
+                    dataset_metadata = json.load(metadata_file)
+            except Exception as e:
+                sys.exit("Error: Unable to read metadata: " + repr(e))
+    elif type(metadata) is dict:
+        dataset_metadata = metadata
+    else:
+        sys.exit("Error: Invalid metadata parameter")
+
+
+    dataset_validator = Validator(dataset_metadata)
+
+
+    # Get the data
+    for data_file in tqdm(find_files(input_path, "^OUTCAR$"), desc="Processing files", disable= not verbose):
+        data = parse_ase(os.path.join(data_file["path"], data_file["filename"]), "vasp")
+        record_metadata = {
+            "mdf-title": dataset_metadata["mdf-title"] + " - " + data["chemical_formula"],
+#            "mdf-acl": ,
+
+            "mdf-tags": ["dft"],
+#            "mdf-description": ,
+            
+            "mdf-composition": data["chemical_formula"],
+#            "mdf-raw": ,
+
+            "mdf-links": {
+#                "mdf-landing_page": ,
+
+#                "mdf-publication": ,
+#                "mdf-dataset_doi": ,
+
+#                "mdf-related_id": ,
+
+                "outcar": {
+ 
+                    "globus_endpoint": "82f1b5c6-6e9b-11e5-ba47-22000b92c6ec",
+                    "http_host": "https://data.materialsdatafacility.org",
+
+                    "path": "/collections/" + data_file["no_root_path"] + "/" + data_file["filename"],
+                    },
+                },
+
+#            "mdf-citation": ,
+#            "mdf-data_contact": {
+
+#                "given_name": ,
+#                "family_name": ,
+
+#                "email": ,
+#                "institution":,
+
+                # IDs
+#                },
+
+#            "mdf-author": ,
+
+#            "mdf-license": ,
+#            "mdf-collection": ,
+            "mdf-data_format": "vasp",
+            "mdf-data_type": "dft",
+#            "mdf-year": ,
+
+#            "mdf-mrr":
+
+#            "mdf-processing": ,
+#            "mdf-structure":,
+            }
+
+        # Pass each individual record to the Validator
+        result = dataset_validator.write_record(record_metadata)
+
+        # Check if the Validator accepted the record, and print a message if it didn't
+        # If the Validator returns "success" == True, the record was written successfully
+        if result["success"] is not True:
+            print("Error:", result["message"])
+
+
+    if verbose:
+        print("Finished converting")
